Removed description.html from upstream repo structure
I suggest to take out the requirement of having a description html snippet in the upstream repository. Indeed, it is often re-written downstream and we want to avoid conflicts between files with the packager. A descriptive README.md is much more important, and we can build the description from it.